Navigation
- added card interaction handlers,
- removed unnecessary box "components"
