+++ deezer-client/src/main/java/components/cards/PlaylistCard.java
@@ -1,6 +1,6 @@
-package components.containers.cards;
+package components.cards;
 
-import api.objects.playables.Artist;
+import api.objects.playables.Playlist;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.control.Button;
@@ -12,24 +12,24 @@ import javafx.scene.layout.VBox;
 import java.io.IOException;
 import java.util.ResourceBundle;
 
-public class ArtistCard extends VBox {
+public class PlaylistCard extends VBox {
 
     @FXML
     private ResourceBundle resources;
     @FXML
-    private Button artistRedirectButton;
+    private Button playlistRedirectButton;
     @FXML
     private ImageView picture;
     @FXML
-    private Label name;
+    private Label title;
     @FXML
-    private Label followers;
+    private Label trackCount;
 
     private Runnable action = () -> {};
 
-    public ArtistCard() {
+    public PlaylistCard() {
         ResourceBundle bundle = ResourceBundle.getBundle("localisation/localisation");
-        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("artistCard.fxml"), bundle);
+        FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource("playlistCard.fxml"), bundle);
         fxmlLoader.setRoot(this);
         fxmlLoader.setController(this);
         try {
@@ -39,23 +39,23 @@ public class ArtistCard extends VBox {
         }
     }
 
-    @FXML
-    private void onArtistRedirection() {
-        action.run();
-    }
-
     public void setAction(Runnable action) {
         this.action = action;
     }
 
-    public void setArtist(Artist artist) {
-        picture.setImage(new Image(artist.getPicture_medium().toString(), true));
+    public void setPlaylist(Playlist playlist) {
+        picture.setImage(new Image(playlist.getPicture_medium().toString(), true));
         picture.fitWidthProperty().bind(this.prefWidthProperty());
         picture.fitHeightProperty().bind(picture.fitWidthProperty());
-        artistRedirectButton.prefWidthProperty().bind(picture.fitWidthProperty());
-        artistRedirectButton.prefHeightProperty().bind(picture.fitHeightProperty());
-        name.setText(artist.getName());
-        followers.setText(String.format("%s: %s", resources.getString("followers"), artist.getNb_fan()));
-        followers.setVisible(artist.getNb_fan() > 0);
+        playlistRedirectButton.prefWidthProperty().bind(picture.fitWidthProperty());
+        playlistRedirectButton.prefHeightProperty().bind(picture.fitHeightProperty());
+        title.setText(playlist.getTitle());
+        trackCount.setText(String.format("%s: %d", resources.getString("tracksCnt"),
+                playlist.getNb_tracks()));
+    }
+
+    @FXML
+    private void onPlaylistRedirection() {
+        action.run();
     }
 }
