--- deezer-client/src/main/java/components/containers/boxes/ArtistBox.java
@@ -1,51 +1,0 @@
-package components.containers.boxes;
-
-import api.objects.playables.Artist;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.HBox;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
-
-import java.util.List;
-
-public class ArtistBox extends VBox {
-    public void fill(List<Artist> artists) {
-        getChildren().clear();
-        for (final Artist artist : artists) {
-            HBox artistBox = new HBox();
-            artistBox.setPadding(new Insets(0, 10, 0, 0));
-            artistBox.setPrefHeight(Region.USE_COMPUTED_SIZE);
-            artistBox.setPrefWidth(Region.USE_COMPUTED_SIZE);
-
-            ImageView artistPicture = new ImageView(new Image(artist.getPicture_small().toString(), true));
-            artistPicture.setFitWidth(56);
-            artistPicture.setFitHeight(56);
-            Button artistBtn = new Button(null, artistPicture);
-            artistBtn.getStyleClass().add("deezer-button");
-            artistBox.getChildren().addAll(artistBtn);
-
-            VBox artistInfoBox = new VBox();
-            artistInfoBox.setPadding(new Insets(10, 0, 0, 0));
-            artistInfoBox.setAlignment(Pos.CENTER);
-            artistInfoBox.setPrefHeight(Region.USE_COMPUTED_SIZE);
-            artistInfoBox.setPrefWidth(Region.USE_COMPUTED_SIZE);
-            artistBox.getChildren().add(artistInfoBox);
-
-            Label artistName = new Label(artist.getName());
-            artistName.getStyleClass().add("deezer-secondary");
-            artistInfoBox.getChildren().add(artistName);
-
-            /*Label artistFollowers = new Label(String.format("%s: %d",
-                    resources.getString("followers"), artist.getNb_fan()));
-            artistFollowers.getStyleClass().add("deezer-secondary");
-            artistInfoBox.getChildren().add(artistFollowers);*/
-
-            getChildren().add(artistBox);
-        }
-    }
-}
