+++ deezer-client/src/main/java/controllers/AlbumPageController.java
@@ -5,8 +5,8 @@ import api.PartialSearchResponse;
 import api.objects.playables.Album;
 import api.objects.playables.Artist;
 import api.objects.playables.TrackSearch;
-import components.containers.cards.AlbumCard;
-import components.containers.cards.ArtistCard;
+import components.cards.AlbumCard;
+import components.cards.ArtistCard;
 import javafx.beans.binding.Bindings;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
