+++ deezer-client/src/main/java/controllers/HomePageController.java
@@ -6,9 +6,9 @@ import api.events.handlers.DeezerListener;
 import api.objects.playables.Album;
 import api.objects.playables.Artist;
 import api.objects.playables.Playlist;
-import components.containers.cards.AlbumCard;
-import components.containers.cards.ArtistCard;
-import components.containers.cards.PlaylistCard;
+import components.cards.AlbumCard;
+import components.cards.ArtistCard;
+import components.cards.PlaylistCard;
 import javafx.beans.binding.Bindings;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
