+++ deezer-client/src/main/java/controllers/SearchPageController.java
@@ -7,11 +7,11 @@ import api.objects.playables.Playlist;
 import api.objects.playables.TrackSearch;
 import api.objects.utils.User;
 import api.objects.utils.search.FullSearchSet;
-import components.containers.cards.AlbumCard;
-import components.containers.cards.ArtistCard;
-import components.containers.cards.PlaylistCard;
-import components.containers.cards.UserCard;
-import components.containers.tables.TrackTable;
+import components.cards.AlbumCard;
+import components.cards.ArtistCard;
+import components.cards.PlaylistCard;
+import components.cards.UserCard;
+import components.tables.TrackTable;
 import javafx.beans.binding.Bindings;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
