+++ deezer-client/src/main/java/controllers/UserPageController.java
@@ -8,10 +8,10 @@ import api.objects.playables.Playlist;
 import api.objects.playables.Track;
 import api.objects.utils.User;
 import api.objects.utils.search.SearchOrder;
-import components.containers.cards.AlbumCard;
-import components.containers.cards.ArtistCard;
-import components.containers.cards.PlaylistCard;
-import components.containers.tables.TrackTable;
+import components.cards.AlbumCard;
+import components.cards.ArtistCard;
+import components.cards.PlaylistCard;
+import components.tables.TrackTable;
 import javafx.beans.binding.Bindings;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
