+++ deezer-client/src/main/resources/components/cards/commentCard.fxml
@@ -5,11 +5,11 @@
 <?import javafx.scene.image.*?>
 <?import javafx.scene.layout.*?>
 
-<fx:root type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1">
+<fx:root type="javafx.scene.layout.HBox" xmlns:fx="http://javafx.com/fxml/1">
    <padding>
       <Insets left="24.0" right="24.0" />
    </padding>
-    <Button fx:id="userRedirectButton" styleClass="deezer-button">
+    <Button fx:id="userRedirectButton" styleClass="deezer-button" onAction="#onUserRedirection">
         <graphic>
             <ImageView fx:id="commentorImg" fitHeight="30.0" fitWidth="30.0" />
         </graphic>
