+++ deezer-client/src/main/resources/components/cards/playlistCard.fxml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.image.*?>
+<?import javafx.scene.layout.*?>
+
+<fx:root maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" styleClass="playlistcard" type="javafx.scene.layout.VBox" xmlns:fx="http://javafx.com/fxml/1">
+    <Button fx:id="playlistRedirectButton"
+            maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity"
+            styleClass="playlistcard-playlist"
+            onAction="#onPlaylistRedirection">
+        <graphic>
+            <ImageView fx:id="picture" />
+        </graphic>
+    </Button>
+    <Label fx:id="title" styleClass="playlistcard-title" wrapText="true" />
+    <Label fx:id="trackCount" styleClass="playlistcard-tracks" wrapText="true" />
+</fx:root>
