+++ deezer-client/src/main/resources/components/tables/trackTable.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.*?>
 
 
-<TableView fixedCellSize="28.0" xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1">
+<TableView fixedCellSize="28.0" xmlns:fx="http://javafx.com/fxml/1">
    <columns>
       <TableColumn fx:id="idxCol" prefWidth="75.0" text="Column X" />
       <TableColumn prefWidth="75.0" text="Column X" />
