+++ deezer-client/src/main/resources/fxml/artistPage.fxml
@@ -1,8 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import components.containers.boxes.ArtistBox?>
-<?import components.containers.boxes.CommentBox?>
-<?import components.containers.boxes.PlaylistBox?>
 <?import javafx.geometry.*?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.image.*?>
@@ -46,7 +43,7 @@
                         <Button fx:id="artistPlaylistsBtn" mnemonicParsing="false" onAction="#artistPlaylistsBtn_OnAction" styleClass="deezer-button" text="%playlists" GridPane.columnIndex="1" />
                         <Button fx:id="artistMtPReleaseBtn" mnemonicParsing="false" styleClass="deezer-button" text="%mostPopularRelease" GridPane.rowIndex="2" />
                         <Button fx:id="artistSimiliarBtn" mnemonicParsing="false" onAction="#artistSimiliarBtn_OnAction" styleClass="deezer-button" text="%artistRelated" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-                        <PlaylistBox fx:id="artistPlaylistsBox" GridPane.columnIndex="1" GridPane.rowIndex="1" GridPane.vgrow="ALWAYS" />
+                        <VBox fx:id="artistPlaylistsBox" GridPane.columnIndex="1" GridPane.rowIndex="1" GridPane.vgrow="ALWAYS" />
                         <TableView fx:id="artistTopTracksTV" maxHeight="100.0" GridPane.rowIndex="1">
                             <columns>
                                 <TableColumn fx:id="artistTMPTIdxCol" prefWidth="75.0" text="%number" />
@@ -59,7 +56,7 @@
                                 <Insets left="40.0" right="40.0" />
                             </GridPane.margin>
                         </TableView>
-                        <ArtistBox fx:id="artistRelatedBox" GridPane.columnIndex="1" GridPane.rowIndex="3" GridPane.vgrow="ALWAYS" />
+                        <VBox fx:id="artistRelatedBox" GridPane.columnIndex="1" GridPane.rowIndex="3" GridPane.vgrow="ALWAYS" />
                         <HBox spacing="20.0" GridPane.rowIndex="3">
                             <VBox HBox.hgrow="NEVER">
                                 <ImageView fx:id="artistTopAlbumImg" fitHeight="200.0" fitWidth="200.0" pickOnBounds="true" preserveRatio="true" />
@@ -172,7 +169,7 @@
                             <Insets top="45.0" />
                         </VBox.margin>
                     </Label>
-                    <CommentBox fx:id="artistCommentsBox" VBox.vgrow="ALWAYS" />
+                    <VBox fx:id="artistCommentsBox" VBox.vgrow="ALWAYS" />
                 </VBox>
                 <padding>
                     <Insets left="80.0" right="80.0" />
