+++ deezer-client/src/main/resources/fxml/playlistPage.fxml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<?import javafx.scene.image.ImageView?>
 <?import javafx.geometry.Insets?>
-<?import components.containers.boxes.CommentBox?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.*?>
 <VBox xmlns:fx="http://javafx.com/fxml" alignment="CENTER" spacing="25.0" fx:controller="controllers.PlaylistPageController">
     <HBox spacing="20.0">
         <VBox.margin>
@@ -68,5 +66,5 @@
         </contextMenu>
     </TableView>
     <Label styleClass="deezer-title" text="%comments" />
-    <CommentBox fx:id="playlistCommentariesBox" prefHeight="200.0" prefWidth="100.0" styleClass="card" />
+    <VBox fx:id="playlistCommentariesBox" prefHeight="200.0" prefWidth="100.0" styleClass="card" />
 </VBox>
