+++ deezer-client/src/main/resources/fxml/searchPage.fxml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import components.containers.tables.TrackTable?>
+<?import components.tables.TrackTable?>
 <?import javafx.geometry.*?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
